requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pycryptodome==3.19.0
cryptography==41.0.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
hiredis==2.3.2
django-redis==5.4.0
python-dotenv==1.0.0
requests==2.31.0
